feat : 장바구니, fix axios

+++ server/services/categories.js
@@ -26,34 +26,46 @@ export default class CategoriesService {
   }
 
   //새로운 카테고리 등록
-  static async addCategory({ title }) {
+  static async addCategory({ title, role }) {
     // 추가 작업 : 이미 등록된 카테고리 여부 체크
-    const createCategory = await Category.create({ title });
-    return createCategory;
+    if (role === 'ADMIN') {
+      const createCategory = await Category.create({ title });
+      return createCategory;
+    } else {
+      throw new Error('권한이 없습니다.');
+    }
   }
 
   //id로 카테고리 수정하기
-  static async updateCategoryById(id, title) {
-    const updateCategory = await Category.findByIdAndUpdate(
-      id,
-      { title },
-      {
-        new: true,
-      },
-    );
-    if (updateCategory) {
-      throw new Error('해당 카테고리는 없습니다.');
+  static async updateCategoryById({ id, title, role }) {
+    if (role === 'ADMIN') {
+      const updateCategory = await Category.findByIdAndUpdate(
+        id,
+        { title },
+        {
+          new: true,
+        },
+      );
+      if (updateCategory) {
+        throw new Error('해당 카테고리는 없습니다.');
+      }
+      return updateCategory;
+    } else {
+      throw new Error('권한이 없습니다');
     }
-    return updateCategory;
   }
 
   //id로 카테고리 삭제하기
-  static async deleteCategoryById(id) {
-    const deletedCategory = await Category.findByIdAndDelete(id);
-    if (!deletedCategory) {
-      throw new Error('해당 카테고리는 없습니다.');
+  static async deleteCategoryById(id, role) {
+    if (role === 'ADMIN') {
+      const deletedCategory = await Category.findByIdAndDelete(id);
+      if (!deletedCategory) {
+        throw new Error('해당 카테고리는 없습니다.');
+      }
+      return deletedCategory;
+    } else {
+      throw new Error('권한이 없습니다.');
     }
-    return deletedCategory;
   }
 
   //id로 상품 가져오기
